Add 'OIP' to major in model definition
The changes in this commit add 'OIP' to the 'major' attribute of the 'parcours_s3' dictionary. Previously, 'OIP' was appended to the 'major' in the 'parcours_ue' dictionary, meaning it was only associated with the major after the filtering of UE parcours. This commit fixes that by ensuring 'OIP' is included in the major definition from the outset.